% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_if_prefix_is_supplemental}
\alias{check_if_prefix_is_supplemental}
\title{Check if a label prefix is a supplemental figure}
\usage{
check_if_prefix_is_supplemental(prefix)
}
\arguments{
\item{prefix}{\code{character} string with caption label prefix.}
}
\value{
\code{boolean}.
}
\description{
Check if Label ends with \code{"S"} or contains \code{"Supplemental"}. Case
insensitive.
}

