% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_marker_type}
\alias{extract_marker_type}
\title{Extract marker type from marker name.}
\usage{
extract_marker_type(marker_name)
}
\arguments{
\item{marker_name}{a \code{character} string with the new marker name.}
}
\value{
\code{character} string containing the marker type.
}
\description{
Extract marker type from marker name.
}

