% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_type_count}
\alias{get_type_count}
\title{Count the number of markers with a specific marker type}
\usage{
get_type_count(marker_obj, type)
}
\arguments{
\item{marker_obj}{\code{marker_obj}.}

\item{type}{\code{character} string with marker type to count.}
}
\value{
\code{integer} with the number of markers of type \code{type}.
}
\description{
Count the number of markers with a specific marker type
}

