% Generated by roxygen2 (4.0.2): do not edit by hand
\name{label}
\alias{label}
\title{Functions to create and reference table and figure numbers.}
\usage{
label(marker_obj, marker_name)
}
\arguments{
\item{marker_obj}{Named \code{list} with caption marker and numbers.}

\item{marker_name}{\code{character} string with marker name. marker name
should have the format <type>.<marker> where <type> is the figure or
table type. For example \code{"fig.car_plot"} or
\code{"tab.p_values"}.}
}
\value{
Named \code{list} with new marker and number
}
\description{
Takes a named list of markers and their number.
Adds \code{marker_name} to \code{marker_obj} with proper number.
The \code{marker_name} is parsed to determine the marker type, i.e. figure
or table based on the marker prefix.
}
\details{
Each marker prefix is given a seperate counter.
}
\seealso{
\code{\link{extract_marker_type}}
}

