% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prefix_ends_in_s}
\alias{prefix_ends_in_s}
\title{Check if prefix ends in S.}
\usage{
prefix_ends_in_s(prefix)
}
\arguments{
\item{prefix}{\code{character} string with caption prefix.}
}
\value{
\code{boolean}
}
\description{
Check if prefix ends in S.
}

