% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_caption}
\alias{print_caption}
\title{Print caption with caption label.}
\usage{
print_caption(marker_obj, marker, prefix, caption)
}
\arguments{
\item{marker_obj}{\code{marker_obj} containing markers and their labels.}

\item{marker}{\code{character} string with marker name.}

\item{prefix}{\code{character} string with caption label prefix.}

\item{caption}{\code{character} string with caption text.}
}
\value{
\code{character} string with full caption.
}
\description{
Print caption with caption label.
}

