% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_label}
\alias{print_label}
\title{For printing labels in a caption or other text environment.}
\usage{
print_label(marker_obj, marker, prefix)
}
\arguments{
\item{marker_obj}{\code{list} containing the markers and their numbers.}

\item{marker}{\code{character}. Name of the marker to print.}

\item{prefix}{\code{character} string with label prefix.}
}
\value{
\code{character} string with complete caption label
}
\description{
For printing labels in a caption or other text environment.
}
\seealso{
\code{\link{check_if_prefix_is_supplemental}}
}

