% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ref}
\alias{ref}
\title{Reference a predefined marker and print the assigned number.}
\usage{
ref(marker_obj, marker)
}
\arguments{
\item{marker_obj}{Named \code{list} containing markers and their numbers.}

\item{marker}{\code{character} string of marker to reference.}
}
\value{
\code{character} containing marker number or "??" for undefined
 reference.
}
\description{
Reference a predefined marker and print the assigned number.
}

