% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarize_markers}
\alias{summarize_markers}
\title{Create a table of all markers used in the document.}
\usage{
summarize_markers(marker_obj)
}
\arguments{
\item{marker_obj}{A \code{list} containing all markers.}
}
\value{
\code{data.frame} with marker name, marker type, and label number.
}
\description{
Create a table that lists all markers by type and number. The columns include
\code{c("Marker", "Type", "Label")}
}

